deps: update application-service/cdq-analysis
The devfile/library module made a breaking change [1] between v2.2.1 and
v2.2.2.

The latest version of application-service depends on devfile/library
v2.2.2. Build-service depends on an older version of application-service
that depended on v2.2.1. Because of the breaking change, it is now
impossible to use the latest application-service together with the
latest build-service.

Update to the latest application-service in order to transitively update
devfile/library to v2.2.2.

[1]: https://github.com/devfile/library/commit/ab4c336ba882ccf675f690ed5435b2a236cdb9dc#diff-9af22c824f5afe3181c1680d45a1702fd3b92bed9d4032e2e9563fd95e844374L60-R64

Signed-off-by: Adam Cmiel <acmiel@redhat.com>